Fixed issue with number parsing.
Fixed jq statement used to parse the number of minutes ago a store was
checked. Due to a recent change on mcbroken.com's end, number of minutes
ago parsing would fail since the numerical field is no longer always in
the second word of the .properties.last_checked field. Replace fixed
offset parsing with a regex match. This is slower, but more correct.