Update retry logic in workflow
Wrap curl + jq command with retry action allowing for both steps to be
retried if there's an issue parsing the downloaded content. Most issues
seem to occur in jq now that we added retries to curl. So now if there's
a failure in parsing the downloaded content, we re-download it and try
again (up to 3 times).